+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Enhancements
 
+- GLS-34 - Improve product classification tree rendering
 - GLS-85 - remove heading from homepage
 - GLS-80 - migration to opensearch
 - GLS-9 - Improve accessibility of learning content reveals